requirements.txt
openai>=1.0.0
fastapi>=0.115.0
uvicorn>=0.32.0
python-multipart>=0.0.20
jinja2>=3.1.0
fastapi>=0.115.0
uvicorn>=0.32.0
python-multipart>=0.0.20
streamlit>=1.39.0
python-dotenv>=1.0.0